requirements.txt
streamlit
rembg[cpu]
pillow
numpy
opencv-python-headless